Make the release build on the platforms it claims to build on
Three separate reasons the first tagged run produced nothing:

`FactoryFlags` is `c_int` on Windows and `c_uint` everywhere else, so
`flags & kUnicode` only type-checks on Windows. Widening both to i64
compares them without a cfg split, and without a cast that is a no-op on
one platform and a lint on the other.

The clap-host fixture tests load clap-test-plugin's cdylib, which cargo
does not build for another package's tests — `cargo test` alone fails on
a clean target directory, and passed here only because a previous build
had left the artifact behind. memo/CHECKS.md already says to build first;
the workflow now does what the file says.

macos-13 is no longer picked up by a runner, so that leg sat queued
rather than failing. macos-15-intel is the Intel runner that still
exists.

Verified by cross-checking both --all-targets for x86_64-unknown-linux-gnu
and aarch64-apple-darwin, and by deleting the fixture cdylib to reproduce
the two Windows test failures and confirm the added build step fixes them.

+++ crates/vst3-host/src/module.rs
@@ -222,7 +222,10 @@ impl Module {
             vendor: from_char8(&raw.vendor),
             url: from_char8(&raw.url),
             email: from_char8(&raw.email),
-            unicode: raw.flags & kUnicode != 0,
+            // `FactoryFlags` は Windows で `c_int`、他で `c_uint` と符号が違う。
+            // どちらからも損なわずに広げられる i64 で突き合わせれば、
+            // プラットフォームごとの分岐もキャストの重複も要らない。
+            unicode: raw.flags as i64 & kUnicode as i64 != 0,
         })
     }
